+++ daemon/src/main/java/resources/logback.xml
