+++ settings.gradle
@@ -3,12 +3,12 @@ include 'common'
 include 'p2p'
 include 'core'
 include 'cli'
+include 'daemon'
 include 'desktop'
 include 'monitor'
 include 'pricenode'
 include 'relay'
 include 'seednode'
 include 'statsnode'
-include 'grpc'
 
 rootProject.name = 'bisq'
